feat: Create Fairpost singleton, add logging there
